* 'develop' of https://github.com/nuls-io/nuls_2.0:
  websocket
  modify chain manage tx logic
  websocket
  remove older rpc call。

+++ chain/nuls-chain/src/main/java/io/nuls/chain/storage/AssetStorage.java
@@ -43,7 +43,7 @@ public interface AssetStorage {
      * @param chainId Chain ID
      * @return List of asset
      */
-    List<Asset> getByChain(short chainId);
+    List<Asset> getByChain(int chainId);
 
     /**
      * Get asset by symbol
